requirements.txt
torch
torchvision
pygame
numpy
pyyaml
matplotlib